requirements.txt
# Recomendadas para desarrollo local. No es obligatorio ejecutar Airflow en Windows.
pyspark==3.4.1
pandas==2.2.2
numpy
pytest==7.4.0
google-cloud-bigquery==3.11.0
apache-airflow==2.6.3
python-dateutil==2.8.2
faker==18.9.0
